+++ packages/react-magma-dom/src/components/Select/Select.test.js
@@ -3,6 +3,7 @@ import { render, fireEvent, act } from '@testing-library/react';
 import { Select } from '.';
 import { defaultI18n } from '../../i18n/default';
 import { magma } from '../../theme/magma';
+import { Modal } from '../Modal';
 
 describe('Select', () => {
   const labelText = 'Label';
@@ -380,6 +381,33 @@ describe('Select', () => {
     });
 
     expect(getByText(items[0])).toBeInTheDocument();
+  });
+
+  it('should close the menu when escape key is pressed, and retain the active modal', () => {
+    const { getByLabelText, getByText, getByTestId, queryByText } = render(
+      <Modal testId="modal" isOpen>
+        <Select labelText={labelText} items={items} />
+      </Modal>
+    );
+
+    const renderedSelect = getByLabelText(labelText, { selector: 'div' });
+
+    renderedSelect.focus();
+
+    fireEvent.keyDown(renderedSelect, {
+      key: ' ',
+    });
+
+    expect(getByText(items[0])).toBeInTheDocument();
+
+    fireEvent.keyDown(getByLabelText(labelText, { selector: 'div' }), {
+      key: 'Escape',
+      code: 27,
+    });
+
+    expect(queryByText(items[0], { selector: 'Red' })).not.toBeInTheDocument();
+
+    expect(getByTestId('modal')).toBeInTheDocument();
   });
 
   it('should not open select when clicking another key other than the enter or spacebar', () => {