+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 5.0.4-next.3
+
+### Patch Changes
+
+- Updated dependencies [8f1bcbe31]
+  - react-magma-dom@4.2.0-next.3
+
+## 5.0.4-next.2
+
+### Patch Changes
+
+- Updated dependencies [9f2aa435d]
+  - react-magma-dom@4.2.0-next.2
+
+## 5.0.4-next.1
+
+### Patch Changes
+
+- Updated dependencies [b5f6189fa]
+  - react-magma-dom@4.2.0-next.1
+
 ## 5.0.4-next.0
 
 ### Patch Changes